% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configureTrial.R
\name{configureTrial}
\alias{configureTrial}
\title{Convert calls into quantile functions}
\usage{
configureTrial(..., var_prefix)
}
\arguments{
\item{...}{calls describing the marginal distributions of endpoints. See \code{...}
in \code{simulateTrialData}.}

\item{var_prefix}{prefix to name variables that users do not give to those.}
}
\value{
a list of quantile functions.
}
\description{
Calls come from ... in \code{simulateTrialData}. For individual level data,
empirical quantile will be generated using \code{simdata::quantile_functions_from_data}.
For calls specifying common distribution families and their parameters in plain text,
quantile functions will be generated by quantile functions in \code{stats}, e.g.,
qnorm, qunif, etc. For calls of user-customized quantile functions, it returns
those directly.
}
