% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df.R
\docType{data}
\name{df}
\alias{df}
\title{Example data from \code{simdata} package}
\format{
A data frame with 4709 rows and 7 variables:
\describe{
\item{Gender}{Gender}
\item{Age}{Age}
\item{Race}{Race}
\item{Weight}{Weight}
\item{BMI}{BMI}
\item{BPsys}{BPsys}
\item{BPdia}{BPdia}
}
}
\source{
{simdata} R package.
}
\usage{
data(df)
}
\description{
A data frame containing the variables as follows:
}
\examples{
data(df)
df

}
\keyword{datasets}
