% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.easySimData.R
\name{plot.easySimData}
\alias{plot.easySimData}
\title{Plot data generated by the \code{easySimData} package.}
\usage{
\method{plot}{easySimData}(x, ...)
}
\arguments{
\item{x}{an object of class \code{easySimData} returned by \code{simulateTrialData}.}

\item{...}{ignored.}
}
\value{
an object returned by \code{gridExtra::grid.arrange}, a grid of plots
comparing the simulated marginal distribution with the ones specified by
users.
}
\description{
Plot data generated by the \code{easySimData} package.
}
